Minor fixes
I've made some minor fixes. I've also added my own R website to the "resources" section. I encourage you to take a look to it and decide whether to add it or not. 

Suggestion: add dots to the end of all phrases or delete all for consistency.

Thanks in advance.